created unit tests and fixed module naming
