+++ test/trivial_config.json
@@ -1,0 +1,2 @@
+{"email" : {"to": ["toemail@gmail.com"], "from": "api@mydomain.com", "smtp_server": "smtp.gmail.com", "smtp_port": "smtp.gmail.com", "password": "FakePassword"},
+"file_system" : {"working_dir": "path/to/file", "remote_dir": "/images/", "dropbox_upload": "path/to/dropboxuploader/or/use/command/if/executable/sytem-wide/dropbox_uploader.sh"}}
