+++ timelapsepi/sample-configure.json
